haystack_app/requirements.txt
flask==2.3.2
farm-haystack==1.16.1
farm-haystack[elasticsearch]==1.16.1
farm-haystack[preprocessing]==1.16.1